requirements.txt
# NEXUS - Neural EXperience Unified System
# Python Dependencies

# Core Framework
streamlit>=1.28.0
psutil>=5.9.0
pyyaml>=6.0.1

# AI & Machine Learning
mlx-lm>=0.2.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Data Processing & Analytics
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# System Integration
psutil>=5.9.0
pyobjc-framework-SystemConfiguration>=9.2
pyobjc-framework-Cocoa>=9.2
pyobjc-framework-Quartz>=9.2

# Web & API
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration & Environment
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
tabulate>=0.9.0
